requirements.txt
streamlit
spacy
pandas
numpy
scikit-learn
pdfminer.six
matplotlib
reportlab
google-generativeai
python-docx